simple_auth/backend-django/requirements/local.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile requirements/local.in
#
aiohttp==3.8.5
    # via
    #   -r requirements/project.txt
    #   aiohttp-retry
    #   twilio
aiohttp-retry==2.8.3
    # via
    #   -r requirements/project.txt
    #   twilio
aiosignal==1.3.1
    # via
    #   -r requirements/project.txt
    #   aiohttp
asgiref==3.7.2
    # via
    #   -r requirements/project.txt
    #   django
async-timeout==4.0.2
    # via
    #   -r requirements/project.txt
    #   aiohttp
attrs==23.1.0
    # via
    #   -r requirements/project.txt
    #   aiohttp
    #   jsonschema
    #   referencing
certifi==2023.7.22
    # via
    #   -r requirements/project.txt
    #   requests
cffi==1.15.1
    # via
    #   -r requirements/project.txt
    #   cryptography
charset-normalizer==3.2.0
    # via
    #   -r requirements/project.txt
    #   aiohttp
    #   requests
cryptography==41.0.3
    # via
    #   -r requirements/project.txt
    #   social-auth-core
defusedxml==0.7.1
    # via
    #   -r requirements/project.txt
    #   python3-openid
    #   social-auth-core
django==4.2.4
    # via
    #   -r requirements/project.txt
    #   django-df-auth
    #   django-otp
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   drf-spectacular
    #   drf-spectacular-sidecar
    #   social-auth-app-django
django-df-auth==0.28
    # via -r requirements/project.txt
django-environ==0.10.0
    # via -r requirements/project.txt
django-otp==1.2.2
    # via
    #   -r requirements/project.txt
    #   django-df-auth
    #   django-otp-twilio
django-otp-twilio==1.0.4
    # via
    #   -r requirements/project.txt
    #   django-df-auth
django-rest-framework==0.1.0
    # via
    #   -r requirements/project.txt
    #   django-df-auth
djangorestframework==3.14.0
    # via
    #   -r requirements/project.txt
    #   django-rest-framework
    #   djangorestframework-simplejwt
    #   drf-spectacular
djangorestframework-simplejwt==5.2.2
    # via
    #   -r requirements/project.txt
    #   django-df-auth
drf-spectacular[sidecar]==0.26.4
    # via -r requirements/project.txt
drf-spectacular-sidecar==2023.8.1
    # via
    #   -r requirements/project.txt
    #   drf-spectacular
frozenlist==1.4.0
    # via
    #   -r requirements/project.txt
    #   aiohttp
    #   aiosignal
idna==3.4
    # via
    #   -r requirements/project.txt
    #   requests
    #   yarl
inflection==0.5.1
    # via
    #   -r requirements/project.txt
    #   drf-spectacular
jsonschema==4.19.0
    # via
    #   -r requirements/project.txt
    #   drf-spectacular
jsonschema-specifications==2023.7.1
    # via
    #   -r requirements/project.txt
    #   jsonschema
multidict==6.0.4
    # via
    #   -r requirements/project.txt
    #   aiohttp
    #   yarl
oauthlib==3.2.2
    # via
    #   -r requirements/project.txt
    #   requests-oauthlib
    #   social-auth-core
pycparser==2.21
    # via
    #   -r requirements/project.txt
    #   cffi
pyjwt==2.8.0
    # via
    #   -r requirements/project.txt
    #   djangorestframework-simplejwt
    #   social-auth-core
    #   twilio
python3-openid==3.2.0
    # via
    #   -r requirements/project.txt
    #   social-auth-core
pytz==2023.3
    # via
    #   -r requirements/project.txt
    #   djangorestframework
    #   twilio
pyyaml==6.0.1
    # via
    #   -r requirements/project.txt
    #   drf-spectacular
referencing==0.30.2
    # via
    #   -r requirements/project.txt
    #   jsonschema
    #   jsonschema-specifications
requests==2.31.0
    # via
    #   -r requirements/project.txt
    #   django-otp-twilio
    #   requests-oauthlib
    #   social-auth-core
    #   twilio
requests-oauthlib==1.3.1
    # via
    #   -r requirements/project.txt
    #   social-auth-core
rpds-py==0.9.2
    # via
    #   -r requirements/project.txt
    #   jsonschema
    #   referencing
social-auth-app-django==5.2.0
    # via
    #   -r requirements/project.txt
    #   django-df-auth
social-auth-core==4.4.2
    # via
    #   -r requirements/project.txt
    #   social-auth-app-django
sqlparse==0.4.4
    # via
    #   -r requirements/project.txt
    #   django
twilio==8.5.0
    # via
    #   -r requirements/project.txt
    #   django-df-auth
typing-extensions==4.7.1
    # via
    #   -r requirements/project.txt
    #   asgiref
uritemplate==4.1.1
    # via
    #   -r requirements/project.txt
    #   drf-spectacular
urllib3==2.0.4
    # via
    #   -r requirements/project.txt
    #   requests
yarl==1.9.2
    # via
    #   -r requirements/project.txt
    #   aiohttp
